refactor reporting tests
