+++ lib/aggregation/reporting/src/test/json/consumer-4.json
@@ -1,0 +1,196 @@
+{
+  "id": "k/610f6508-8b5d-4840-888d-0615ade33117/4ef2f706-f2ae-4be5-a18c-40a969cf8fb6/UNKNOWN/t/1448457444188",
+  "consumer_id": "UNKNOWN",
+  "organization_id": "610f6508-8b5d-4840-888d-0615ade33117",
+  "resource_instance_id": "0b39fa70-a65f-4183-bae8-385633ca5c87",
+  "start": 1448284898000,
+  "end": 1448457443000,
+  "resources": [
+    {
+      "resource_id": "test-resource-id",
+      "plans": [
+        {
+          "plan_id": "basic/test-metering-plan/test-rating-plan/test-pricing-basic",
+          "metering_plan_id": "test-metering-plan",
+          "rating_plan_id": "test-rating-plan",
+          "pricing_plan_id": "test-pricing-basic",
+          "aggregated_usage": [
+            {
+              "metric": "memory",
+              "windows": [
+                [
+                  null
+                ],
+                [
+                  {
+                    "quantity": {
+                      "consumed": 0,
+                      "consuming": 0.5
+                    },
+                    "cost": {
+                      "consumed": 0,
+                      "consuming": 0.5,
+                      "price": 0.00014
+                    },
+                    "summary": 0.036679722222222225,
+                    "charge": 0.00000513516111111111
+                  }
+                ],
+                [
+                  {
+                    "quantity": {
+                      "consumed": 156250,
+                      "consuming": 0.625
+                    },
+                    "cost": {
+                      "consumed": 156250,
+                      "consuming": 0.625,
+                      "price": 0.00014
+                    },
+                    "summary": 0.08925243055555555,
+                    "charge": 0.00001249534027777778
+                  }
+                ],
+                [
+                  {
+                    "quantity": {
+                      "consumed": 19684375,
+                      "consuming": 7.125
+                    },
+                    "cost": {
+                      "consumed": 19684375,
+                      "consuming": 7.125,
+                      "price": 0.00014
+                    },
+                    "summary": 5.990567986111111,
+                    "charge": 0.0008386795180555555
+                  }
+                ],
+                [
+                  {
+                    "quantity": {
+                      "consumed": 1155809375,
+                      "consuming": 7.125
+                    },
+                    "cost": {
+                      "consumed": 1155809375,
+                      "consuming": 7.125,
+                      "price": 0.00014
+                    },
+                    "summary": 321.5808457638889,
+                    "charge": 0.04502131840694444
+                  }
+                ]
+              ]
+            }
+          ],
+          "windows": [
+            [
+              null
+            ],
+            [
+              {
+                "charge": 0.00000513516111111111
+              }
+            ],
+            [
+              {
+                "charge": 0.00001249534027777778
+              }
+            ],
+            [
+              {
+                "charge": 0.0008386795180555555
+              }
+            ],
+            [
+              {
+                "charge": 0.04502131840694444
+              }
+            ]
+          ]
+        }
+      ],
+      "aggregated_usage": [
+        {
+          "metric": "memory",
+          "windows": [
+            [
+              null
+            ],
+            [
+              {
+                "charge": 0.00000513516111111111
+              }
+            ],
+            [
+              {
+                "charge": 0.00001249534027777778
+              }
+            ],
+            [
+              {
+                "charge": 0.0008386795180555555
+              }
+            ],
+            [
+              {
+                "charge": 0.04502131840694444
+              }
+            ]
+          ]
+        }
+      ],
+      "windows": [
+        [
+          null
+        ],
+        [
+          {
+            "charge": 0.00000513516111111111
+          }
+        ],
+        [
+          {
+            "charge": 0.00001249534027777778
+          }
+        ],
+        [
+          {
+            "charge": 0.0008386795180555555
+          }
+        ],
+        [
+          {
+            "charge": 0.04502131840694444
+          }
+        ]
+      ]
+    }
+  ],
+  "windows": [
+    [
+      null
+    ],
+    [
+      {
+        "charge": 0.00000513516111111111
+      }
+    ],
+    [
+      {
+        "charge": 0.00001249534027777778
+      }
+    ],
+    [
+      {
+        "charge": 0.0008386795180555555
+      }
+    ],
+    [
+      {
+        "charge": 0.04502131840694444
+      }
+    ]
+  ]
+}
